backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
python-multipart==0.0.9
openpyxl==3.1.2
lxml==5.2.2
python-Levenshtein==0.25.1
pydantic==2.7.1
pydantic-settings==2.2.1
